refactoring and cleanup of daily_traffic_provider.dart
